prism/nuke_license_parser/pyproject.toml
[project]
name = "nuke-license-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "pyside6>=6.11.0",
    "python-dotenv>=1.2.2",
    "qtpy>=2.4.3",
    "requests>=2.33.1",
]
